[events] Events for Bndtools
This patch instruments bnd for sending out
events that can be used by bndtools. This patch
does not yet use the events but it verifies that all 
is the same.

Next step is to use the events.

+++ bndtools.builder/src/org/bndtools/builder/MarkerSupport.java
@@ -91,8 +91,10 @@ class MarkerSupport {
 			deleteMarkers(MARKER_BND_MISSING_WORKSPACE);
 			deleteMarkers(MARKER_JAVA_BASELINE);
 		} else {
-			project.deleteMarkers(markerType, true, IResource.DEPTH_INFINITE);
-			deleteCnfMarkersForThisProject(markerType);
+			if (project.exists() && project.isAccessible()) {
+				project.deleteMarkers(markerType, true, IResource.DEPTH_INFINITE);
+				deleteCnfMarkersForThisProject(markerType);
+			}
 		}
 	}
